remade new dataset backend
Overall, added a bunch of the backend to allow the user to create datasets.

Forms.py:
Added a temporary form NewDatasetForm to use a placeholder till the front-end form is created.
new_dataset.html:
Added a temporary html document till the front-end html is created.
file_downloader.py:
Created a new file and methods to create a temporary file using a URL and otherstuff. The HTTPS downloader is implemented. The SFTP downloader is still a work in progress.

schema_generator.py:
created a file and placeholder method for the schema generator till phase 3.

Views.py:
Created a new view new_dataset which takes in the NewDataSetForm and attmempts to create a schema from the given file or url.