Added code to parse different file types
schema_generator.py:
The schema_generator function was changed to take a file name and chooses which function to call to generate the schema.
The csv_schema_generator method now reads the first line and splits it into seperate pieces which should be the metadata for the file.
The xlsx_schema_generator method creates a worksheet using the openpyxl library and gets the metadata from the first row in the first sheet.
views.py:
An inputted file is now checked to make sure it matches the supported file types. Also, the file name or url (which would end with the file name) is now submitted in the schema_generator method.